added python venv to setup.sh

+++ pi-code/setup.sh
@@ -1,3 +1,10 @@
 sudo apt update
-sudo apt install python-serial
-pip install -r requirements.txt
+sudo apt install python3-serial
+
+python3 -m venv .venv
+source .venv/bin/activate
+if [ -f "requirements.txt" ]; then
+    pip install -r requirements.txt
+else
+    echo "requirements.txt not found!"
+fi
